Move sample reads download endpoint to `caches.api`
