+++ virtool/samples/api.py
@@ -1100,41 +1100,3 @@ async def download_artifact(req: aiohttp.web.Request):
     }
 
     return FileResponse(file_path, chunk_size=1024 * 1024, headers=headers)
-
-
-@routes.jobs_api.get("/api/samples/{sample_id}/caches/{key}/reads/reads_{suffix}.fq.gz")
-async def download_reads_cache(req):
-    """
-    Download sample reads cache for a given key.
-
-    """
-    db = req.app["db"]
-    pg = req.app["pg"]
-
-    sample_id = req.match_info["sample_id"]
-    key = req.match_info["key"]
-    suffix = req.match_info["suffix"]
-
-    file_name = f"reads_{suffix}.fq.gz"
-
-    if not await db.samples.count_documents({"_id": sample_id}) or not await db.caches.count_documents({"key": key}):
-        return not_found()
-
-    existing_reads = await virtool.samples.files.get_existing_reads(pg, sample_id, key=key)
-
-    if file_name not in existing_reads:
-        return not_found()
-
-    file_path = req.app["settings"]["data_path"] / "caches" / key / file_name
-
-    if not os.path.isfile(file_path):
-        return not_found()
-
-    file_stats = virtool.utils.file_stats(file_path)
-
-    headers = {
-        "Content-Length": file_stats["size"],
-        "Content-Type": "application/gzip"
-    }
-
-    return FileResponse(file_path, chunk_size=1024 * 1024, headers=headers)
